+++ src/backend/manager.rs
@@ -189,7 +189,7 @@ impl Manager {
     ) -> Result<impl Iterator<Item = Content>, ApplicationError> {
         crate::trace!("Querying message by thread: {:?}, from {:?}", thread, from);
         if let Some(config_store) = self.imp().config_store.borrow().as_ref() {
-            Ok(config_store.messages(thread, from)?)
+            Ok(config_store.messages(thread, from)?.filter_map(|o| o.ok()))
         } else {
             log::error!("Query messages by contact without config store being set up");
             // TODO: Error?